pyproject.toml
[tool.poetry]
name = "fabrique_notification_service"
version = "0.1.0"
description = ""
authors = ["Stanis96 <vasilievstas96@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.6"
djangorestframework = "^3.13.1"
django-environ = "^0.9.0"
drf-yasg = "^1.21.3"
celery = "^5.2.7"
redis = "^4.3.4"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
flake8 = "^5.0.0"
pyproject-flake8 = "^0.0.1-alpha.4"
flake8-black = "^0.3.3"
flake8-builtins = "^1.5.3"
flake8-import-order = "^0.18.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
exclude = """
    ('
    |.idea/
    |.venv/
    |.vscode/
    ')
"""

# call using pflake8
[tool.flake8]
max-line-length = 100
max-complexity = 8
ignore = "A003, W503"
exclude = """
    .idea/
    .venv/
    .vscode/
    */migrations/
"""

[virtualenvs]
create = true
in-project = true
path = ".venv"
